requirements.txt
bcrypt==4.2.0
gunicorn
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
ffmpeg-python==0.2.0
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-mail==0.10.0
flask-sqlalchemy==3.1.1
future==1.0.0
greenlet==3.0.3
idna==3.7
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mutagen==1.47.0
pycryptodomex==3.20.0
PyJWT==2.9.0
requests==2.32.3
SQLAlchemy==2.0.32
typing-extensions==4.12.2
urllib3==2.2.2
websockets==12.0
werkzeug==3.0.3
youtube-dl==2021.12.17
yt-dlp==2024.8.6
zipp==3.19.2